refactor(pi-fence): move FenceMode to types.ts
